extract EditorHelper to component proj
